+++ bzt/modules/grinder.py
@@ -24,7 +24,7 @@ from bzt.engine import ScenarioExecutor, Scenario, FileLister, PythonGenerator
 from bzt.modules.aggregator import ConsolidatingAggregator, ResultsReader
 from bzt.modules.console import WidgetProvider, ExecutorWidget
 from bzt.six import iteritems
-from bzt.utils import shell_exec, MirrorsManager, dehumanize_time
+from bzt.utils import shell_exec, MirrorsManager, dehumanize_time, get_full_path
 from bzt.utils import unzip, RequiredTool, JavaVM, shutdown_process, TclLibrary
 
 
@@ -356,13 +356,12 @@ class Grinder(RequiredTool):
             return False
 
     def install(self):
-        dest = os.path.dirname(os.path.dirname(os.path.expanduser(self.tool_path)))
-        dest = os.path.abspath(dest)
-        grinder_dist = super(Grinder, self).install_with_mirrors(dest, ".zip")
-        self.log.info("Unzipping %s", grinder_dist.name)
-        unzip(grinder_dist.name, dest, 'grinder-' + self.version)
-        grinder_dist.close()
-        os.remove(grinder_dist.name)
+        dest = get_full_path(self.tool_path, step_up=2)
+        self.log.info("Will install %s into %s", self.tool_name, dest)
+        grinder_dist = self._download()
+        self.log.info("Unzipping %s", grinder_dist)
+        unzip(grinder_dist, dest, 'grinder-' + self.version)
+        os.remove(grinder_dist)
         self.log.info("Installed grinder successfully")
         if not self.check_if_installed():
             raise RuntimeError("Unable to run %s after installation!" % self.tool_name)