+++ virtualenv-test.bash
@@ -2,10 +2,6 @@
 
 rm -rf build/tools
 
-# setup node.js
-rm -rf node_modules
-npm install mocha
-
 # setup env
 virtualenv --clear build
 source build/bin/activate