+++ site/dat/docs/Changelog.md
@@ -9,6 +9,10 @@
  - fix `timeout` handling in Locust script generator
  - fix Locust crash when `concurrency` isn't specified
  - support scenario-level `headers` in Locust script generator
+ - add exception throw to tools downloader
+ - add Mocha autoinstallation for Selenium tests
+ - add `selenium-webdriver` npm package installation for Mocha runner
+ - make `-locations` respect `use-deprecated-api` flag
 
 ## 1.7.0 <sup>2 oct 2016</sup>
  - add RSpec tests runner for Selenium